regroup settings layout component in one file
