+++ src/components/layout/NavMenu/NavMenu.tsx
@@ -94,7 +94,7 @@ export const NavMenu = ({ teams, user }: Props) => {
                       {teams && teams.length > 0 && <Divider />}
                       {!!currentTeam && (
                         <Item
-                          href={routes.TEAMS_EDIT.replace(
+                          href={routes.TEAM_EDIT_PROFILE.replace(
                             ':slug',
                             currentTeam!.slug
                           )}
